% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{knox}
\alias{knox}
\title{Original Knox test}
\usage{
knox(x, y, time, sigma, tau)
}
\arguments{
\item{x}{Longitude, should be projected to a planar system.}

\item{y}{Latitude, should be projected to a planar system.}

\item{time}{time column.}

\item{sigma}{a distance in space.}

\item{tau}{a distance in time.}
}
\value{
A knox statistic, which is the number of pairs of points found in a given space-time distance.
}
\description{
Original Knox test
}
\references{
\strong{Knox, E. (1964)}. \emph{The detection of space-time interactions.} Journal of the Royal Statistical Society. Series C (13(1), 25-30.

\strong{Tango, T. (2010)}. \emph{Statistical methods for disease clustering. Springer.}
}

