% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{knox_mc}
\alias{knox_mc}
\title{Knox test with Monte Carlo simulation}
\usage{
knox_mc(x, y, time, sigma, tau, perm = 999L)
}
\arguments{
\item{x}{Longitude, should be projected to a planar system.}

\item{y}{Latitude, should be projected to a planar system.}

\item{time}{time column.}

\item{sigma}{a distance in space.}

\item{tau}{a distance in time.}

\item{perm}{Number of simulations. Default is 999}
}
\value{
\item{knox}{Knox statistic, which is the number of pairs of points found in a given space-time distance.}
\item{p_value}{p-value calculated from simulation}
\item{RR}{Relative Risk - calculated by observed value (Knox statistics) divided by mean of simulated values}
}
\description{
Performing Knox statistics with Monte Carlo simulation
}

