% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_st_link.R
\name{writeSHP}
\alias{writeSHP}
\title{Writting shapefile of space-time links}
\usage{
writeSHP(data, x, y, time, ds, dt, w = TRUE, file.out = "st_links", ...)
}
\arguments{
\item{data}{A data frame holds at least 3 columns: longtitude, latitude and
time.}

\item{x}{Longitude, should be projected to a planar system.}

\item{y}{Latitude, should be projected to a planar system.}

\item{time}{time column.}

\item{ds}{a cut-off distance in space.}

\item{dt}{a cut-off distance in time.}

\item{w}{default is TRUE, to write out a shapefile.}

\item{file.out}{default is "st_links", shapfile name to be written.}

\item{...}{parameters of the plot function of a line, i.e col, width, etc.}
}
\value{
A plot of space-time links in lines. By default, \code{w} is set
  TRUE, a shapefile will be written out as well.
}
\description{
Writing shapefile from a data frame of space-time links
}
\author{
Thanh Le Viet, \email{lethanhx2k@gmail.com}
}
\seealso{
\code{\link{st_links}}
}

